update plot tests
