+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-addopts = --mpl-baseline-path=tests/data/mpl
+addopts = --mpl-baseline-path=tests/data/mpl --mpl
 testpaths = tests
