+++ tests/bmds3/models/test_dichotmous.py
@@ -67,6 +67,12 @@ class TestBmdModelDichotomous:
         assert not np.isclose(resp1.bmd, resp2.bmd)
         assert resp1.bmd < resp2.bmd
 
+    @pytest.mark.mpl_image_compare
+    def test_bmds3_dichotomous(self, ddataset2):
+        model = dichotomous.Logistic(dataset=ddataset2)
+        model.execute()
+        return model.plot()
+
 
 @pytest.mark.skipif(not RunBmds3.should_run, reason=RunBmds3.skip_reason)
 def test_bmds3_dichotomous_models(ddataset2):
