+++ tests/data/cassettes/tests.bmds2.test_plots/test_dichotomous_hill_plot.yaml
