+++ tests/data/cassettes/tests.bmds2.test_plots/test_exponential_m2_plot.yaml
