+++ tests/data/cassettes/tests.bmds2.test_plots/test_exponential_m4_plot.yaml
