+++ tests/data/cassettes/tests.bmds2.test_plots/test_gamma_plot.yaml
