+++ tests/data/cassettes/tests.bmds2.test_plots/test_linear_plot.yaml
