+++ tests/data/cassettes/tests.bmds2.test_plots/test_logistic_plot.yaml
