+++ tests/data/cassettes/tests.bmds2.test_plots/test_loglogistic_plot.yaml
