+++ tests/data/cassettes/tests.bmds2.test_plots/test_logprobit_plot.yaml
