+++ tests/data/cassettes/tests.bmds2.test_plots/test_multistage_cancer_plot.yaml
