+++ tests/data/cassettes/tests.bmds2.test_plots/test_multistage_plot.yaml
