+++ tests/data/cassettes/tests.bmds2.test_plots/test_polynomial_plot.yaml
