+++ tests/data/cassettes/tests.bmds2.test_plots/test_power_plot.yaml
