+++ tests/data/cassettes/tests.bmds2.test_plots/test_probit_plot.yaml
