+++ tests/data/cassettes/tests.bmds2.test_plots/test_weibull_plot.yaml
